new GroupExamineSystem
